requirements.txt
streamlit
plotly
pandas
bcrypt==4.0.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1

bcrypt
pandas
plotly
psycopg2-binary
python-dotenv
streamlit
